third_party/blink/web_tests/http/tests/devtools/portals/portals-console-expected.txt

Tests that the console works correctly with portals

Running: testMainConsole
!!window.portalHost = false

Running: testPortalConsole
!!window.portalHost = true
